+++ data/glsk-file/src/main/java/com/farao_community/farao/commons/data/glsk_file/EICode.java
@@ -1,14 +1,15 @@
 /**
- * Copyright (c) 2018, RTE (http://www.rte-france.com)
+ * Copyright (c) 2019, RTE (http://www.rte-france.com)
  * This Source Code Form is subject to the terms of the Mozilla Public
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
  */
-package com.farao_community.farao.commons.data.generatorloadshiftkeys;
+package com.farao_community.farao.commons.data.glsk_file;
 
 import com.powsybl.iidm.network.Country;
 
 /**
+ * @author RTE International {@literal <contact@rte-international.com>}
  * EIC = Energy Identification Code
  */
 public class EICode {
