+++ data/glsk-file/src/main/java/com/farao_community/farao/commons/data/glsk_file/actors/GlskDocumentImporter.java
@@ -1,12 +1,12 @@
 /**
- * Copyright (c) 2018, RTE (http://www.rte-france.com)
+ * Copyright (c) 2019, RTE (http://www.rte-france.com)
  * This Source Code Form is subject to the terms of the Mozilla Public
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
  */
-package com.farao_community.farao.commons.data.generatorloadshiftkeys.actors;
+package com.farao_community.farao.commons.data.glsk_file.actors;
 
-import com.farao_community.farao.commons.data.generatorloadshiftkeys.GlskDocument;
+import com.farao_community.farao.commons.data.glsk_file.GlskDocument;
 import org.xml.sax.SAXException;
 
 import javax.xml.parsers.ParserConfigurationException;
@@ -19,6 +19,7 @@ import java.nio.file.Path;
 /**
  * Importer for CIM type GlskDocument
  * Create a GlskDocument object
+ * @author RTE International {@literal <contact@rte-international.com>}
  */
 public class GlskDocumentImporter {
