+++ data/glsk-file/src/main/java/com/farao_community/farao/commons/data/glsk_file/actors/GlskPointLinearGlskConverter.java
@@ -1,12 +1,12 @@
 /**
- * Copyright (c) 2018, RTE (http://www.rte-france.com)
+ * Copyright (c) 2019, RTE (http://www.rte-france.com)
  * This Source Code Form is subject to the terms of the Mozilla Public
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
  */
-package com.farao_community.farao.commons.data.generatorloadshiftkeys.actors;
+package com.farao_community.farao.commons.data.glsk_file.actors;
 
-import com.farao_community.farao.commons.data.generatorloadshiftkeys.*;
+import com.farao_community.farao.commons.data.glsk_file.*;
 import com.powsybl.iidm.network.Country;
 import com.powsybl.iidm.network.Generator;
 import com.powsybl.iidm.network.Load;
@@ -24,6 +24,7 @@ import java.util.stream.Collectors;
 
 /**
  * Convert a single GlskPoint to LinearGlsk
+ * @author RTE International {@literal <contact@rte-international.com>}
  */
 public class GlskPointLinearGlskConverter {
