+++ data/glsk-file/src/main/java/com/farao_community/farao/commons/data/glsk_file/actors/UcteGlskDocumentScalableConverter.java
@@ -1,15 +1,15 @@
 /**
- * Copyright (c) 2018, RTE (http://www.rte-france.com)
+ * Copyright (c) 2019, RTE (http://www.rte-france.com)
  * This Source Code Form is subject to the terms of the Mozilla Public
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
  */
-package com.farao_community.farao.commons.data.generatorloadshiftkeys.actors;
+package com.farao_community.farao.commons.data.glsk_file.actors;
 
 import com.farao_community.farao.commons.chronology.DataChronology;
 import com.farao_community.farao.commons.chronology.DataChronologyImpl;
-import com.farao_community.farao.commons.data.generatorloadshiftkeys.GlskPoint;
-import com.farao_community.farao.commons.data.generatorloadshiftkeys.UcteGlskDocument;
+import com.farao_community.farao.commons.data.glsk_file.GlskPoint;
+import com.farao_community.farao.commons.data.glsk_file.UcteGlskDocument;
 import com.powsybl.action.util.Scalable;
 import com.powsybl.iidm.network.Network;
 import org.xml.sax.SAXException;
@@ -28,6 +28,7 @@ import java.util.Map;
  * return a map:
  * Key: country
  * Value: DataChronology of Scalable
+ * @author RTE International {@literal <contact@rte-international.com>}
  */
 public class UcteGlskDocumentScalableConverter {
