+++ data/glsk-file/src/test/java/com/farao_community/farao/commons/data/glsk_file/GlskDocumentLinearGlskConverterTest.java
@@ -1,14 +1,14 @@
 /**
- * Copyright (c) 2018, RTE (http://www.rte-france.com)
+ * Copyright (c) 2019, RTE (http://www.rte-france.com)
  * This Source Code Form is subject to the terms of the Mozilla Public
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
  */
-package com.farao_community.farao.commons.data.generatorloadshiftkeys;
+package com.farao_community.farao.commons.data.glsk_file;
 
 import com.farao_community.farao.commons.chronology.DataChronology;
-import com.farao_community.farao.commons.data.generatorloadshiftkeys.actors.GlskDocumentImporter;
-import com.farao_community.farao.commons.data.generatorloadshiftkeys.actors.GlskDocumentLinearGlskConverter;
+import com.farao_community.farao.commons.data.glsk_file.actors.GlskDocumentImporter;
+import com.farao_community.farao.commons.data.glsk_file.actors.GlskDocumentLinearGlskConverter;
 import com.powsybl.iidm.import_.Importers;
 import com.powsybl.iidm.network.Network;
 import com.powsybl.sensitivity.factors.variables.LinearGlsk;
@@ -28,6 +28,9 @@ import java.util.Map;
 import static org.junit.Assert.assertFalse;
 import static org.junit.Assert.assertTrue;
 
+/**
+ * @author RTE International {@literal <contact@rte-international.com>}
+ */
 public class GlskDocumentLinearGlskConverterTest {
     private static final Logger LOGGER = LoggerFactory.getLogger(GlskDocumentLinearGlskConverterTest.class);
