+++ data/glsk-file/src/test/java/com/farao_community/farao/commons/data/glsk_file/GlskPointScalableConverterTest.java
@@ -1,13 +1,13 @@
 /**
- * Copyright (c) 2018, RTE (http://www.rte-france.com)
+ * Copyright (c) 2019, RTE (http://www.rte-france.com)
  * This Source Code Form is subject to the terms of the Mozilla Public
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
  */
-package com.farao_community.farao.commons.data.generatorloadshiftkeys;
+package com.farao_community.farao.commons.data.glsk_file;
 
-import com.farao_community.farao.commons.data.generatorloadshiftkeys.actors.GlskDocumentImporter;
-import com.farao_community.farao.commons.data.generatorloadshiftkeys.actors.GlskPointScalableConverter;
+import com.farao_community.farao.commons.data.glsk_file.actors.GlskDocumentImporter;
+import com.farao_community.farao.commons.data.glsk_file.actors.GlskPointScalableConverter;
 import com.google.common.math.DoubleMath;
 import com.powsybl.action.util.Scalable;
 import com.powsybl.iidm.import_.Importers;
@@ -23,6 +23,9 @@ import java.io.IOException;
 
 import static org.junit.Assert.assertTrue;
 
+/**
+ * @author RTE International {@literal <contact@rte-international.com>}
+ */
 public class GlskPointScalableConverterTest {
     private static final Logger LOGGER = LoggerFactory.getLogger(GlskPointScalableConverterTest.class);
     private static final String GLSKB42TEST = "/GlskB42test.xml";
