+++ data/glsk-file/src/test/java/com/farao_community/farao/commons/data/glsk_file/UcteGlskDocumentImporterTest.java
@@ -1,12 +1,12 @@
 /**
- * Copyright (c) 2018, RTE (http://www.rte-france.com)
+ * Copyright (c) 2019, RTE (http://www.rte-france.com)
  * This Source Code Form is subject to the terms of the Mozilla Public
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
  */
-package com.farao_community.farao.commons.data.generatorloadshiftkeys;
+package com.farao_community.farao.commons.data.glsk_file;
 
-import com.farao_community.farao.commons.data.generatorloadshiftkeys.actors.UcteGlskDocumentImporter;
+import com.farao_community.farao.commons.data.glsk_file.actors.UcteGlskDocumentImporter;
 import com.google.common.math.DoubleMath;
 import org.junit.Test;
 import org.slf4j.Logger;
@@ -21,6 +21,9 @@ import java.util.Map;
 
 import static org.junit.Assert.*;
 
+/**
+ * @author RTE International {@literal <contact@rte-international.com>}
+ */
 public class UcteGlskDocumentImporterTest {
     private static final Logger LOGGER = LoggerFactory.getLogger(UcteGlskDocumentImporterTest.class);
