+++ data/glsk-file/src/test/java/com/farao_community/farao/commons/data/glsk_file/UcteGlskDocumentLinearGlskConverterTest.java
@@ -1,13 +1,13 @@
 /**
- * Copyright (c) 2018, RTE (http://www.rte-france.com)
+ * Copyright (c) 2019, RTE (http://www.rte-france.com)
  * This Source Code Form is subject to the terms of the Mozilla Public
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
  */
-package com.farao_community.farao.commons.data.generatorloadshiftkeys;
+package com.farao_community.farao.commons.data.glsk_file;
 
 import com.farao_community.farao.commons.chronology.DataChronology;
-import com.farao_community.farao.commons.data.generatorloadshiftkeys.actors.UcteGlskDocumentLinearGlskConverter;
+import com.farao_community.farao.commons.data.glsk_file.actors.UcteGlskDocumentLinearGlskConverter;
 import com.powsybl.iidm.import_.Importers;
 import com.powsybl.iidm.network.Network;
 import com.powsybl.sensitivity.factors.variables.LinearGlsk;
@@ -25,6 +25,9 @@ import java.util.Map;
 import static org.junit.Assert.assertFalse;
 import static org.junit.Assert.assertTrue;
 
+/**
+ * @author RTE International {@literal <contact@rte-international.com>}
+ */
 public class UcteGlskDocumentLinearGlskConverterTest {
 
     private static final String UCTETEST = "/20170322_1844_SN3_FR2_GLSK_test.xml";
